# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"schemas": {
			"Task": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"_v": {
						"type": "number"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				}
			},
			"Image": {
				"properties": {
					"task": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"imagePath": {
						"type": "string"
					},
					"width": {
						"type": "string"
					},
					"contetType": {
						"type": "string"
					},
					"_v": {
						"type": "number"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				}
			}
		}
	},
	"info": {
		"title": "bcncgroup-image-test",
		"version": "1.0.0",
		"description": "Resize images system",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Alberto Castillo"
		}
	},
	"servers": [
		{
			"url": "http://localhost:8089",
			"description": "Local server"
		}
	],
	"tags": [
		{
			"name": "Example Routes",
			"description": "Hello World"
		},
		{
			"name": "Task"
		},
		{
			"name": "Images"
		}
	],
	"paths": {
		"/": {
			"get": {
				"summary": "Hello World example route",
				"description": "",
				"operationId": "helloWorld",
				"tags": [
					"Example Routes"
				],
				"responses": {
					"200": {
						"description": "Succesfully get Hello World message",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Hello World!!"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"statusCode": {
											"type": "number",
											"example": 500
										},
										"name": {
											"type": "string",
											"example": "Internal Server Error"
										},
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/v1/image/download/{taskId}/{width}": {
			"get": {
				"operationId": "DownloadImage",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"image/png": {},
							"image/jpeg": {}
						}
					}
				},
				"tags": [
					"Images"
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "width",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/image/downloadById/{id}": {
			"get": {
				"operationId": "DownloadImageById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Images"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/image/images": {
			"get": {
				"operationId": "GetImages",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Image"
									}
								},
								"type": "array"
							}
						}
					}
				},
				"tags": [
					"Images"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/task/get/{taskId}": {
			"get": {
				"operationId": "GetTask",
				"tags": [
					"Task"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/task/tasks": {
			"get": {
				"operationId": "GetTasks",
				"tags": [
					"Task"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/task": {
			"post": {
				"operationId": "CreateTask",
				"tags": [
					"Task"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/v1/task/complete/{taskId}": {
			"post": {
				"operationId": "CompleteTask",
				"tags": [
					"Task"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	}
}